% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladdertargets-package.R
\docType{package}
\name{bladdertargets-package}
\alias{bladdertargets}
\alias{bladdertargets-package}
\title{bladdertargets: Data Package For Bladder Cancer Models}
\description{
Contains data to be used by CISNET Bladder Cancer models as targets for calibration and validation. Currently contains the following datasets: (us_pop_40)
}
\author{
\strong{Maintainer}: Andrew Huang \email{andrew_huang2@brown.edu} (https://github.com/aw-huang/bladdertargets)

}
\keyword{internal}
