% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_bystudy_byyear.R
\docType{data}
\name{surv_bystudy_byyear}
\alias{surv_bystudy_byyear}
\title{Surveillance intervals between cystoscopy}
\format{
A tibble with 21 rows and 8 variables:
\describe{
\item{study}{chr Identifier for EORTC trial}
\item{surv_year}{chr }
\item{n_intervals}{num }
\item{days_intervals_mean}{num TBA}
\item{days_intervals_sd}{num TBA}
\item{days_intervals_med}{num TBA}
\item{days_intervals_min}{num TBA}
\item{days_intervals_max}{num TBA}
}
}
\source{
National Intercensal Data Files, U.S. Census Bureau
\url{https://www.census.gov/data/tables/time-series/demo/popest/pre-1980-national.html}
}
\usage{
surv_bystudy_byyear
}
\description{
Based on EORTC trials
}
\keyword{datasets}
